+++ integration/integration_test_commands/integration_test_power_management_vm.py
