+++ tests/test_drivers/test_deployment_driver.py
@@ -3,7 +3,7 @@
 from mock import Mock, MagicMock
 
 from common.cloudshell.data_retriever import *
-from deploymentServiceDriver.DeploymentServiceDriver import *
+from deployment_service.driver import *
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '../vCenterShell/vCenterShell'))
